Add utility class
